back/requirements.txt
Flask==2.2.5
Pillow==10.4.0
torch==2.2.2
torchvision==0.17.2
ultralytics==8.2.66
